Refactored tabs into separate pages
