+++ frontend/src/js/tabs/NavBar.tsx
