+++ frontend/src/js/tabs/metadata/index.tsx
@@ -1,3 +1,3 @@
-import Modal from './Modal';
+import Page from './Page';
 
-export { Modal };
+export default Page;
